OLD/requirements.txt
pycryptodome==3.20.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.3.0
